+++ src/pages/bookshelf/index.ts
@@ -1,4 +1,2 @@
 export { default as BookshelfHeader } from '@/pages/bookshelf/BookshelfHeader.vue'
-export { default as BookshelfNetworkInfo } from '@/pages/bookshelf/BookshelfNetworkInfo.vue'
-export { default as BookshelfPrices } from '@/pages/bookshelf/BookshelfPrices.vue'
 export { default as BookDetails } from '@/pages/bookshelf/BookDetails.vue'